backend/requirements.txt
asgiref==3.2.10
bson==0.5.8
dataclasses==0.7
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djongo==1.3.0
gunicorn==20.0.4
mongoengine==0.20.0
Pillow==7.2.0
pymongo==3.11.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
